+++ R/span_km.R
@@ -27,8 +27,8 @@
 #'
 #' @examples
 #' set.seed(123)
-#' R1 <- matrix(rnorm(300), 100, 3)  # benchmarks: T=100, N=3
-#' R2 <- matrix(rnorm(200), 100, 2)  # tests:      T=100, K=2
+#' R1 <- matrix(rnorm(300), 100, 3)  # benchmarks: T=100, K=3
+#' R2 <- matrix(rnorm(200), 100, 2)  # test assets: T=100, N=2
 #' ans <- span_km(R1, R2)
 #' ans$pval; ans$stat; ans$H0
 #'